Import initial version of ucom
For efficiency, we rely on poll() which has been specified by POSIX since
1994 so should be available on any reasonably recent system, including
Linux and the BSD family.

However, poll() has been broken on macOS for over twenty years: it fails
with POLLNVAL on character devices like terminals, which in turn breaks
ucom. The macOS kernel and command-line userspace are dilapidated and
unmaintained, so it feels unlikely this bug will be fixed any time soon.