Ensure the serial port is in raw mode
The remote end of a serial link will often be in canonical mode with echo
enabled, but the local end should always be in raw mode to avoid loops
and other strange behaviour. Save and restore the port settings in the
same way we do for the local terminal.

Set c_cc[VMIN] = 1 and c_cc[VTIME] = 0 on both the local terminal and
serial port, guarding against systems where these take priority over
O_NONBLOCK on the file descriptors.