Set CLOCAL in termios c_cflag
On BSD kernels, writes to serial tty devices will block if the carrier
line is not asserted, although the corresponding cu devices will not.
For convenience, set the CLOCAL flag to allow tty and cu devices to be
used interchangeably, copying the behaviour of GNU Screen.

Mention the legacy select() branch for macOS in README.