Use select() instead of poll() to work around macOS bugs
On macOS, poll() does not work on character devices, including serial
interfaces and pseudo-terminals. This bug has languished for quarter of a
century, despite the legacy select() interface not being similarly broken.

Competence is unlikely to break out in Apple's kernel team, so adapt
a version of ucom to use select() to work around this mess. On better
operating systems, you probably want the mainline branch of ucom instead.